requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
firebase-admin==6.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
google-generativeai==0.3.2
python-socketio==5.9.0
